refactor(edge): move edge package to internal
