+++ internal/edge/key.go
